% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchitSE.R
\docType{package}
\name{MatchItSE}
\alias{MatchItSE}
\alias{MatchItSE-package}
\title{MatchItSE: methods for Standard Error estimation}
\description{
Contains auxiliary functions for calculating Standard Errors for
matched samples from the MatchIt package. Please note, that Matchit
does not estimate Standard Errors by intention, since matching is seen by the
package’s authors as a data preprocessing step.
}

